@@ -4,10 +4,22 @@
   "description": "Notifications Addon that displays growl-like messages.",
   "main": "dist/index.js",
   "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "prepublish": "npm run build",
-    "test": "npm run build && NODE_ENV=test mocha --compilers js:babel-core/register --require ./test/helpers.js --require ./test/dom.js --recursive",
-    "tdd": "NODE_ENV=test npm test -- --watch"
+    "build": "webpack --config webpack.config.babel.js",
+    "demo": "react-scripts start",
+    "prepare": "npm run build",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "flow": "flow",
+    "lint": "eslint *.js lib",
+    "docs": "jsdoc -c jsdoc.json",
+    "docs:explain": "jsdoc -X -c jsdoc.json",
+    "validate": "npm ls"
+  },
+  "jest": {
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
   },
   "repository": {
     "type": "git",
@@ -23,21 +35,32 @@
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
     "babel-loader": "^7.1.0",
     "babel-preset-es2015": "^6.24.1",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-1": "^6.24.1",
-    "chai": "^4.0.2",
     "enzyme": "^2.9.0",
-    "jsdom": "^9.12.0",
-    "jsdom-global": "^2.1.1",
-    "mocha": "^3.4.2",
+    "enzyme-to-json": "^1.5.1",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.2",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-flowtype": "^2.34.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jest": "^20.0.3",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.1.0",
+    "flow-bin": "^0.49.1",
+    "ink-docstrap": "^1.3.0",
+    "jest": "^20.0.4",
+    "jsdoc": "^3.4.3",
+    "jsdoc-babel": "^0.3.0",
+    "precommit-hook-eslint": "^3.0.0",
     "react": "^15.6.1",
-    "react-addons-test-utils": "^15.6.0",
     "react-dom": "^15.6.1",
     "react-test-renderer": "^15.6.1",
-    "sinon": "^2.3.5",
-    "webpack": "^2.4.1",
+    "webpack": "^3.1.0",
     "webpack-node-externals": "^1.6.0"
   },
   "peerDependencies": {
@@ -46,5 +69,10 @@
   "dependencies": {
     "react-notification-system": "^0.2.14",
     "tproptypes": "^0.1.6"
-  }
+  },
+  "pre-commit": [
+    "lint",
+    "flow",
+    "test"
+  ]
 }
