@@ -1,12 +1,13 @@
+// @flow
+
 import {getNotificationSystem} from './registerNotificationSystem';
 
 /**
  * Displays a notification.
- * @memberOf module:addons/notifications
  * @tag Action
  * @param {Notification} notification - The notification object.
  */
-export function addNotification(notification) {
+export function addNotification(notification: Object) {
 	const nfs = getNotificationSystem();
 	if (nfs) {
 		nfs.addNotification({
@@ -18,11 +19,10 @@ export function addNotification(notification) {
 
 /**
  * Removes a displayed notification
- * @memberOf module:addons/notifications
  * @tag Action
  * @param {Notification} notification - The notification object.
  */
-export function removeNotification(notification) {
+export function removeNotification(notification: Object) {
 	const nfs = getNotificationSystem();
 	if (nfs) nfs.removeNotification(notification);
 }
@@ -30,7 +30,6 @@ export function removeNotification(notification) {
 /**
  * The notification object can have even more properties. See {@link https://github.com/igorprado/react-notification-system#creating-a-notification|React Notification System}
  * for more information.
- * @memberOf module:addons/notifications
  * @typedef Notification
  * @property {string} level - The notification level, 'error', 'success', etc.
  * @property {string} title - The notifcation title.
